refactor: 카드 합계 계산시 ACE 처리 메서드 분리
